Add new pages and custom hook UseRoutes
