+++ client/src/pages/CollectionPage.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const CollectionPage = () => {
+  return (
+    <div>
+      <h1>Collection Page</h1>
+    </div>
+  );
+};
+
+export default CollectionPage;
