+++ client/src/pages/LoginPage.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const LoginPage = () => {
+  return (
+    <div>
+      <h1>Log in Page</h1>
+    </div>
+  );
+};
+
+export default LoginPage;
