+++ client/src/routes.js
@@ -1,0 +1,27 @@
+import React from 'react';
+import { Switch, Route, Redirect } from 'react-router-dom';
+import CollectionPage from './pages/CollectionPage';
+import AuthPage from './pages/AuthPage';
+
+const UseRoutes = (isAuthenticated) => {
+  if (isAuthenticated) {
+    return (
+      <Switch>
+        <Route path="/collection" exact>
+          <CollectionPage />
+        </Route>
+        <Redirect to="/collection" />
+      </Switch>
+    );
+  }
+  return (
+    <Switch>
+      <Route path="/" exact>
+        <AuthPage />
+      </Route>
+      <Redirect to="/" />
+    </Switch>
+  );
+};
+
+export default UseRoutes;
