+++ lib/routes/auth.js
@@ -42,7 +42,9 @@ router.post(
         password: hashedPassword,
       });
       await user.save();
-      response.status(201).json({ message: 'A new user account was created' });
+      response
+        .status(201)
+        .json({ message: 'A new user account has been created' });
     } catch (error) {
       response
         .status(500)
